extract tagger popover logic
