+++ src/renderer/src/components/file/search-annotation/index.tsx
@@ -2,29 +2,11 @@ import type {
   Annotation,
   LabelAnnotation,
 } from "@/components/file-annotator/types";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
+import AnnotationPopover from "@/components/file/annotation-popover";
 import { useAnnotation } from "@/context/Annotation";
-import { css, cva } from "@/styled/css";
+import { cva } from "@/styled/css";
 import type { AllLabels, AllLabelsWithSufix } from "@/types/aymurai";
-import { type FocusEventHandler, useRef, useState } from "react";
 import SearchTagger from "./search-tagger";
-
-const triggerReset = css({
-  appearance: "none",
-  bg: "transparent",
-  border: "[none]",
-  padding: "0",
-  cursor: "default",
-  "&:focus-visible": {
-    outline: "[2px solid token(colors.action.hover)]",
-    outlineOffset: "[2px]",
-    borderRadius: "sm",
-  },
-});
 
 const search = cva({
   base: {
@@ -54,67 +36,23 @@ export default function SearchAnnotation({
 }: SearchAnnotationProps) {
   const { add, addBySearch, createAnnotationData } = useAnnotation();
 
-  const [open, setOpen] = useState(false);
-  const closeTimer = useRef<ReturnType<typeof setTimeout>>(null);
-  const isFocusInside = useRef(false);
-
   const isAnnotable = !!annotateTo;
-
-  const scheduleClose = () => {
-    closeTimer.current = setTimeout(() => {
-      if (!isFocusInside.current) setOpen(false);
-    }, 100);
-  };
-
-  const cancelClose = () => {
-    if (closeTimer.current) clearTimeout(closeTimer.current);
-  };
-
-  const handleMouseEnter = () => {
-    cancelClose();
-    setOpen(true);
-  };
-
-  const handleBlur: FocusEventHandler<HTMLDivElement> = (e) => {
-    if (!e.currentTarget.contains(e.relatedTarget)) {
-      isFocusInside.current = false;
-      scheduleClose();
-    }
-  };
 
   if (!isAnnotable)
     return <mark className={search({ clickable: false })}>{children}</mark>;
 
   return (
-    <>
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger
-          className={triggerReset}
-          onMouseEnter={handleMouseEnter}
-          onMouseLeave={scheduleClose}
-        >
-          <mark className={search({ clickable: true })}>{children}</mark>
-        </PopoverTrigger>
-        <PopoverContent
-          side="top"
-          sideOffset={8}
-          showArrow={false}
-          onMouseEnter={cancelClose}
-          onMouseLeave={scheduleClose}
-          onFocus={() => {
-            isFocusInside.current = true;
-          }}
-          onBlur={handleBlur}
-        >
-          <SearchTagger
-            onLabelChange={console.log}
-            onSuffixChange={console.log}
-            onAddAll={handleAddAll}
-            onAddOne={handleAddOne}
-          />
-        </PopoverContent>
-      </Popover>
-    </>
+    <AnnotationPopover
+      trigger={<mark className={search({ clickable: true })}>{children}</mark>}
+      content={
+        <SearchTagger
+          onLabelChange={console.log}
+          onSuffixChange={console.log}
+          onAddAll={handleAddAll}
+          onAddOne={handleAddOne}
+        />
+      }
+    />
   );
 
   function handleAddOne() {
